refactor(dashboard): extract Mastio key rotation sub-router (F-B-201 PR-10/13)
Move the four /proxy/mastio-key routes (status page + grace-days
config + rotate trigger + complete-staged finalization) plus the
three private helpers (_mastio_rotation_min_interval_seconds,
_load_rotation_grace_days, _mastio_key_to_view) out of the
mcp_proxy/dashboard/router.py god-object into a dedicated
mastio_key_routes.py module, mounted via include_router.

Same pattern as PR-2..PR-9. CSRF gate, approval-hook intercept on
/mastio-key/rotate and the staged-rotation state machine all
preserved verbatim.

F-B-201 progress: 10/13 PR. Three left: users, enrollments+badges,
core leftover.
